Catch up to -current. (CVS: It's not just a program. It's an adventure!)

+++ sys/arch/mipsco/mipsco/disksubr.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: disksubr.c,v 1.8.8.2 2002/02/28 04:10:47 nathanw Exp $	*/
+/*	$NetBSD: disksubr.c,v 1.8.8.3 2002/04/01 07:41:17 nathanw Exp $	*/
 
 /*
  * Copyright (c) 1982, 1986, 1988 Regents of the University of California.
@@ -158,7 +158,7 @@ setdisklabel(olp, nlp, openmask, clp)
 	if (nlp->d_magic != DISKMAGIC || nlp->d_magic2 != DISKMAGIC ||
 	    dkcksum(nlp) != 0)
 		return (EINVAL);
-	while ((i = ffs((long)openmask)) != 0) {
+	while ((i = ffs(openmask)) != 0) {
 		i--;
 		openmask &= ~(1 << i);
 		if (nlp->d_npartitions <= i)
