+++ sys/arch/vax/vax/disksubr.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: disksubr.c,v 1.28.8.2 2002/03/29 23:22:45 ragge Exp $	*/
+/*	$NetBSD: disksubr.c,v 1.28.8.3 2002/04/01 07:43:31 nathanw Exp $	*/
 
 /*
  * Copyright (c) 1982, 1986, 1988 Regents of the University of California.
@@ -158,7 +158,7 @@ setdisklabel(struct disklabel *olp, struct disklabel *nlp,
 	if (nlp->d_magic != DISKMAGIC || nlp->d_magic2 != DISKMAGIC ||
 	    dkcksum(nlp) != 0)
 		return (EINVAL);
-	while ((i = ffs((long)openmask)) != 0) {
+	while ((i = ffs(openmask)) != 0) {
 		i--;
 		openmask &= ~(1 << i);
 		if (nlp->d_npartitions <= i)
